backend/requirements-test.txt
# Testing Dependencies for Django Backend
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
django-test-plus==2.2.2
model-bakery==1.17.0
responses==0.24.1
freezegun==1.2.2
